Add files via upload
PROBLEM STATEMENT & OBJECTIVE
To aid the problem of tracking humans easily & quickly without risking more human lives in desperate times like disasters or terrorist attacks.

“Purpose is to Save Life and 
Save Humanity”

Reasons for loss of life in DISASTERS & TERRORIST ATTACKS.

Not quickly recognizing human life in disaster-prone areas due to this immediate medical treatment not being provided on time.
Also, human involvement requires to search life which requires extra human resources, safety equipment, extra cost & time.

Approach:
The concept of surveillance drones arises to aid the problem of tracking humans quickly without risking more human lives in desperate times like disasters or terrorist attacks. 
The design consists of a Quad-Copter with an object detection camera and GPS to quickly track humans from distance in areas that are not quickly reachable from land and/or are not safe for a human to reach unknowingly. 
The key element here is the detection of humans below rubbles in case of disaster and terrorist attacks. 
Moving Object Detection is a technique used in Artificial Vision. It consists of processing the video so that any object moving within the field of view can be detected. This technique is used to recognize the movement of an object, vehicle, or person through a camera in autonomous mode and display an alarm to the operator.